[feat] 메인페이지 데이터 조회 코드 item.js로 이동
